[ci/msys2] Remove workaround for meson preferring system libraries
The workaround was added in #4354 to avoid the linker using system harfbuzz
over the just built one.

This is/was most likely https://github.com/mesonbuild/meson/pull/12861
which hopefully should be merged upstream for meson 1.8.

In MSYS2 we have backported this fix for over a year since
https://github.com/msys2/MINGW-packages/pull/20058 so using the meson
package provided by MSYS2 (as is the case here) should avoid the problem
already.