[COLRv1] Simplify paint_glyph()
Pass a non-synthetic font to the client, so they don't have
to worry about it.

The only issue left, is that now when paint_glyph() paints
an outline glyph (not color glyph), it would be double-slanted.

I'm leaning towards removing painting of outline glyphs from
the paint API and change the paint API to return a boolean.