[check-*.py] Remove autotoolism
Instead of passing source dir and build dir to as envvars, pass
them directly as script arguments. The scripts also were supposed to be
able to run outsize of the build scripts, but this was broken since the
fallback when the envvars were not set was wrong.

The HBSOURCES and HBHEADERS are still passed as envvars since they are
long lists and might hit command line length limitation on some systems.