Fix cmap4 overflow pathway.
cmap4 serialize was doing a pop_discard() on overflow, however the encoding record above it was also doing one as well. Remove the unnessecary cmap4 pop_dicard(). Note: in cmap4 there was a comment implying the subsequent pop_discard() would be a noop due to the set error, however is not true since pop_discard() still works when there's been an overflow error.