[morx] Speed up unsafe-to-break
This is a port of https://github.com/harfbuzz/harfrust/pull/241
but crashes (with LucidaGrande for example), because, unlike HarfRust,
in HarfBuzz it's not safe to access states up to 64 without checking.

In HarfBuzz we do that check in sanitize() but don't keep the max
state number. Also, we allow negative states.

All considered, looks like this optimization is not worth implementing
in HB, specially since our .entry() is much faster than in HarfRust.